+++ test/unit/lib/context/requestContext.js
@@ -244,6 +244,95 @@ describe('RequestContext', () => {
                     });
             });
 
+            it('should build the correct POST context with Controls.dryRun in the Tenant', () => {
+                const restOp = new RestOperationMock();
+                restOp.method = 'Post';
+                // note extra /
+                restOp.setPath(`${path}/?mock=1`);
+
+                const dryRunDecl = util.simpleCopy(validDecl);
+                dryRunDecl.tenantId.controls = {
+                    class: 'Controls',
+                    dryRun: true
+                };
+                restOp.setBody(dryRunDecl);
+
+                return RequestContext.get(restOp, hostContext)
+                    .then((ctxt) => {
+                        assert.isUndefined(ctxt.error);
+                        assert.strictEqual(ctxt.tasks[0].action, 'deploy');
+                        assert.strictEqual(ctxt.tasks[0].dryRun, true);
+                        assert.strictEqual(ctxt.tasks[0].warnings[0].tenant, 'tenantId');
+                        assert.strictEqual(ctxt.tasks[0].warnings[0].message, 'dryRun true found in Tenant controls');
+                    });
+            });
+
+            it('should build the correct POST context with Controls.dryRun in one of multiple Tenants', () => {
+                const restOp = new RestOperationMock();
+                restOp.method = 'Post';
+                // note extra /
+                restOp.setPath(`${path}/?mock=1`);
+
+                const dryRunDecl = util.simpleCopy(validDecl);
+                dryRunDecl.tenantId.controls = {
+                    class: 'Controls',
+                    dryRun: true
+                };
+                dryRunDecl.tenantId2 = {
+                    class: 'Tenant',
+                    appId: {
+                        class: 'Application'
+                    }
+                };
+                restOp.setBody(dryRunDecl);
+
+                return RequestContext.get(restOp, hostContext)
+                    .then((ctxt) => {
+                        assert.isUndefined(ctxt.error);
+                        assert.strictEqual(ctxt.tasks[0].action, 'deploy');
+                        assert.strictEqual(ctxt.tasks[0].dryRun, true);
+                        assert.strictEqual(ctxt.tasks[0].warnings.length, 1);
+                        assert.strictEqual(ctxt.tasks[0].warnings[0].tenant, 'tenantId');
+                        assert.strictEqual(ctxt.tasks[0].warnings[0].message, 'dryRun true found in Tenant controls');
+                    });
+            });
+
+            it('should build the correct POST context with Controls.dryRun in multiple Tenants with a warning for each Tenant', () => {
+                const restOp = new RestOperationMock();
+                restOp.method = 'Post';
+                // note extra /
+                restOp.setPath(`${path}/?mock=1`);
+
+                const dryRunDecl = util.simpleCopy(validDecl);
+                dryRunDecl.tenantId.controls = {
+                    class: 'Controls',
+                    dryRun: true
+                };
+                dryRunDecl.tenantId2 = {
+                    class: 'Tenant',
+                    appId: {
+                        class: 'Application'
+                    },
+                    controls: {
+                        class: 'Controls',
+                        dryRun: true
+                    }
+                };
+                restOp.setBody(dryRunDecl);
+
+                return RequestContext.get(restOp, hostContext)
+                    .then((ctxt) => {
+                        assert.isUndefined(ctxt.error);
+                        assert.strictEqual(ctxt.tasks[0].action, 'deploy');
+                        assert.strictEqual(ctxt.tasks[0].dryRun, true);
+                        assert.strictEqual(ctxt.tasks[0].warnings.length, 2);
+                        assert.strictEqual(ctxt.tasks[0].warnings[0].tenant, 'tenantId');
+                        assert.strictEqual(ctxt.tasks[0].warnings[0].message, 'dryRun true found in Tenant controls');
+                        assert.strictEqual(ctxt.tasks[0].warnings[1].tenant, 'tenantId2');
+                        assert.strictEqual(ctxt.tasks[0].warnings[1].message, 'dryRun true found in Tenant controls');
+                    });
+            });
+
             it('should build the correct POST context with action dry-run', () => {
                 const restOp = new RestOperationMock();
                 restOp.method = 'Post';
@@ -727,6 +816,78 @@ describe('RequestContext', () => {
                                 }
                             ]
                         );
+                    });
+            });
+
+            it('should build the correct per-app POST context with dry-run in Tenant controls', () => {
+                const restOp = new RestOperationMock();
+                restOp.method = 'Post';
+
+                restOp.setPathName(`${path}/Tenant1/applications/`);
+                restOp.setPath(`${path}/Tenant1/applications/`);
+
+                restOp.setBody({
+                    id: 'per-app-declaration',
+                    schemaVersion: '3.50.0',
+                    controls: {
+                        class: 'Controls',
+                        dryRun: true,
+                        logLevel: 'debug',
+                        trace: true
+                    },
+                    Application1: {
+                        class: 'Application',
+                        service: {
+                            class: 'Service_HTTP',
+                            virtualAddresses: [
+                                '192.0.2.1'
+                            ],
+                            pool: 'pool'
+                        },
+                        pool: {
+                            class: 'Pool',
+                            members: [
+                                {
+                                    servicePort: 80,
+                                    serverAddresses: [
+                                        '192.0.2.10',
+                                        '192.0.2.20'
+                                    ]
+                                }
+                            ]
+                        }
+                    },
+                    Application2: {
+                        class: 'Application',
+                        service: {
+                            class: 'Service_HTTP',
+                            virtualAddresses: [
+                                '192.0.2.2'
+                            ],
+                            pool: 'pool'
+                        },
+                        pool: {
+                            class: 'Pool',
+                            members: [
+                                {
+                                    servicePort: 80,
+                                    serverAddresses: [
+                                        '192.0.2.30',
+                                        '192.0.2.40'
+                                    ]
+                                }
+                            ]
+                        }
+                    }
+                });
+                return RequestContext.get(restOp, hostContext)
+                    .then((ctxt) => {
+                        assert.isUndefined(ctxt.error);
+                        assert.strictEqual(ctxt.tasks[0].action, 'deploy');
+                        assert.strictEqual(ctxt.tasks[0].dryRun, true);
+                        assert.strictEqual(ctxt.tasks[0].warnings.length, 1);
+                        assert.strictEqual(ctxt.tasks[0].warnings[0].tenant, 'Tenant1');
+                        assert.strictEqual(ctxt.tasks[0].warnings[0].message, 'dryRun true found in Tenant controls');
                     });
             });